[01216] Move parseHexAlpha and combineHexAlpha to color-utils.ts
